add cold memory collection and reporting
